requirements.txt
gitpython
anthropic
